Match 87% of `StaffRoll`

+++ include/Game/Camera/CameraContext.hpp
@@ -34,5 +34,5 @@ public:
 };
 
 namespace MR {
-    u16 getScreenHeight();
+    s32 getScreenHeight();
 };
